==> R/01_load.qmd <==
---
title: "Load Data"
output: html
execute:
  echo: true
  output: false
---

#### Loading

Load `GEOquery` library, which is used to retrieve data sets from NCBI's Gene Expression Omnibus(<https://www.bioconductor.org/packages//2.10/bioc/html/GEOquery.html>). Also load `Biobase`, explained further down, and `tidyverse` (<https://tidyverse.tidyverse.org/>).

```{r}
library("GEOquery")
library("Biobase")
library("tidyverse")
```

Fetch data set *GSE11512* and extract first argument according to the GEO data set structure.

```{r}
gset <- getGEO("GSE11512",
               GSEMatrix = TRUE,
               getGPL = FALSE) |>
  pluck(1)
```

#### Saving

Save the data set as an .RDS file in *\_raw.*

```{r}
# Set path for raw data directory and check if exists
raw_data_dir <- file.path("..",
                          "data",
                          "_raw")
if (!dir.exists(raw_data_dir)) {
  dir.create(raw_data_dir,
             recursive = TRUE)
  message("Created directory: ", raw_data_dir)}

saveRDS(gset, file = file.path(raw_data_dir,
                               "GSE11512.RDS"))
```

The raw data consists of two subsets of data that contain the gene expressions and pheno types of samples. These are loaded into each their *dat_load.tsv* file. We use `Biobase` to extract the data of interest, as our data is formatted as an 'ExpressionSet' (<https://www.bioconductor.org/packages//2.10/bioc/html/Biobase.html>).

```{r}
# Set path for processed data directory
processed_data_dir <- file.path("..",
                                "data")

# Write tsv file for expression data
expr_data <- exprs(gset)
write.table(expr_data,
            file = file.path(processed_data_dir,
                             "01.1_dat_load_expr.tsv"),
            sep = "\t",
            row.names = TRUE,
            col.names = NA,
            quote = FALSE)

# Write tsv file for phenotype data
pheno_data <- pData(phenoData(gset))
write.table(pheno_data,
            file = file.path(processed_data_dir,
                             "01.2_dat_load_pheno.tsv"),
            sep = "\t",
            row.names = TRUE,
            col.names = NA,
            quote = TRUE)
```
