==> R/03.2_augment_norm.qmd <==
---
title: "Analysis 1"
output: html
execute:
  echo: true
---

# Normalization of Age

A new dataframe named normalized_joined will be created. Normalized age values will be added to a new column named normalized_age, moved to the front of the dataframe, and rounded to 4 decimal places. Age for all species will be normalized based on average life expectancy from birth found online from the following sources.

-   human: 71.4 years (as of 2021) from [WHO](https://data.who.int/indicators/i/90E2E48)

-   chimpanzee : 32.9 years from [Research Paper](https://pmc.ncbi.nlm.nih.gov/articles/PMC8364750/)

-   macaque: 15 days from [The Macaque Website](https://macaques.nc3rs.org.uk/about-macaques/life-history)

```{r}
normalized_joined <- pheno_expr_data_joined |>
  mutate(
    normalized_age = case_when(
      species == "Human" ~ round(log2((age / 71.4) + 1.1), 4),
      species == "Chimpanzee" ~ round(log2((age / 32.9) + 1.1), 4),
      species == "Macaque" ~ round(log2((age / 15) + 1.1), 4)
    )
  ) |>
  select(normalized_age, everything(), -age)
```
