==> R/04_describe.qmd <==
---
title: "Data Description"
output: html
execute:
  echo: true
  output: false
---

#### Load

Load packages.

```{r}
library("tidyverse")
library("patchwork")
library("ggtext")
```

Load data and color palettes.

```{r}
# Set path for processed data directory
processed_data_dir <- file.path("..",
                                "data")

# Data
pheno_expr_data_joined <- read_tsv(file = file.path(processed_data_dir,
                             "03_dat_joined.tsv"),
                      show_col_types = FALSE)

# Color palette
color_palette <- read_tsv(file = file.path(processed_data_dir,
                                         "09_col_pal.tsv"),
                     show_col_types = FALSE)
```

#### Plot descriptions

Number of observations and variables in the data set.

```{r}
# Number of samples
str_glue("Number of samples {pheno_expr_data_joined |>
         nrow()}")


# Number of variables
str_glue("Number of variables {pheno_expr_data_joined |>
         ncol()}")

# Number of genes
str_glue("Number of genes {pheno_expr_data_joined |>
         ncol() - 4}")
```

### Assigning Colors

Setting up the desired colors for the graphics:

```{r}
species_color <- color_palette |>
  filter(group == "Species") |>
  pull(color)

sex_color <- color_palette |>
  filter(group == "Alternative") |>
  slice_tail(n = 2) |>
  pull(color)
```

### Plotting

Plotting of the number of individuals by sex and species

```{r}
species_plot <- pheno_expr_data_joined |>
  ggplot(mapping = aes(y = species,
                     fill = sex)) +
  geom_bar(color = "black") +
  theme_minimal() +
  theme(
    plot.title.position = "plot",
    legend.position = "none",
    panel.grid.minor = element_blank(),
    panel.grid.major = element_blank(),
    axis.text.y = element_text(size = 12,
                                face = "italic"),
    axis.title = element_blank(),
    plot.subtitle = element_markdown(size = 16
                                     ,)) +
  labs(y = element_blank()) +
  scale_fill_manual(values = color_palette |>
                      filter(group == "Alternative") |>
                      slice_tail(n = 2) |>
                      pull(color))
species_plot

```

Plot distribution of population of data set by age:

```{r}
sex_plot <- pheno_expr_data_joined |>
  ggplot(mapping = aes(x = age,
                       fill = species)) +
  geom_boxplot() +
  theme_minimal() +
  theme(
    plot.title.position = "plot",
    panel.grid.minor = element_blank(),
    panel.grid.major = element_blank(),
    axis.text.y = element_blank(),
    axis.title.x = element_blank(),
    axis.title.y = element_text(size = 12,
                              face = "italic"),
    plot.subtitle = element_markdown(size = 16),
    legend.position = "none") +
  scale_fill_manual(values = color_palette |>
                      filter(group == "Species") |>
                      pull(color))

```

### Combining Plots

```{r}
description_plots <- wrap_elements(
  full =
  gridtext::richtext_grob(
    text = paste0(
      "for <span style='color:",
      species_color |>
        pluck(1),
      ";'>chimpanzees</span>, <span style='color:",
      species_color |>
        pluck(2),
      ";'>humans</span>, and <span style='color:",
      species_color |>
        pluck(3),
      ";'>macaques </span>",
      "by <span style='color:",
      sex_color |> pluck(1),
      ";'>male</span> and ",
      "<span style='color:",
      sex_color |> pluck(2),
      ";'>female</span>."),
    x = 0,
    hjust = -0.05,
    gp = grid::gpar(fontsize = 16))) /
  
  (species_plot + sex_plot +
     
     plot_layout(width = c(4, 6))) +
  
  plot_layout(heights = c(1, 9)) +
  
  plot_annotation(
     title = "Distributions of species (left) and age (right)",
     theme = theme(
       plot.title = element_text(size = 20,
                                 hjust =0)))
  

description_plots
```

#### Saving

Write final plot to *.png* file.

```{r}
ggsave(file = file.path(processed_data_dir,
                        "04_description_plots.png"),
       plot = description_plots,
       width = 12,
       height = 8)
```
