==> R/05_analysis_PCA.qmd <==
---
title: "Principle Component Analysis"
output: html
execute:
  echo: true
  output: false
---

#### Load

For the principal component analysis (PCA) we use `prcomp`, and for the proportion of variation explained (PVE), we use the functions described in *99_proj_func.R*.

```{r}
library("tidyverse")
library("dplyr")
library("broom")
library("patchwork")

source("99_proj_func.R")
```

Load the data set.

```{r}
# Set path for processed data directory
processed_data_dir <- file.path("..",
                                "data")

# Load the dataset
data <- read_tsv(file = file.path(processed_data_dir,
                                         "03_dat_joined.tsv"),
                 show_col_types = FALSE)
```

#### PCA

Numeric encoding of our non-numeric variables of interest and scaling of data.

```{r}

# Numeric encoding and extraction
pca_data <- data |>
  mutate(sex_num = case_when(sex == "M" ~ 1,
                             TRUE ~ 0),
         species_num = case_when(species == "Human" ~ 0,
                                 species == "Chimpanzee" ~ 1,
                                 TRUE ~ 2)) |>
  dplyr::select(where(is.numeric))

# Scale gene expression
pca_data_scaled <- pca_data |>
  scale()
```

Apply the `prcomp()` function and define projected data as tidy format.

```{r}
# Perform PCA
pca_results <- pca_data_scaled |>
  prcomp(scale. = FALSE)

# Tidy projection results
pca_proj_df <- pca_results |> 
  pluck("x") |> 
  as_tibble() |> 
  mutate(species = data |>
           pull(species),
         age = data |>
           pull(age),
         sex = data |>
           pull(sex),
         log_age = pca_data |>
           pull(log_age),
         species_num = pca_data |>
           pull(species_num),
         sex_num = pca_data |>
           pull(sex_num))
```

Extract standard deviation and calculate variance explained and cumulative variance.

```{r}

pca_var_df <- pca_results |> 
  pluck("sdev") |> 
  tibble(sdev = _ ) |>
  mutate(
    PC = paste0("PC",
                row_number()),
    variance_explained = (sdev^2 / sum(sdev^2)) * 100,
    cumulative_variance = cumsum(variance_explained)) |> 
  dplyr::select(PC,
         variance_explained,
         cumulative_variance)
```

#### PVE

Combine data frames by the PC.

```{r}
pca_proj_long_df <- pca_proj_df |>
  pivot_longer(
    cols = starts_with("PC"),
    names_to = "PC",
    values_to = "value")


pca_results_df <- pca_var_df |> 
  left_join(pca_proj_long_df,
            by = "PC")
```

Calculate the variance explained by attribute for each PC.

```{r}
var_explained_pc_all <- pca_results_df |>
  group_by(PC) |>
  nest() |>
  mutate(
    var_explained = map(
      .x = data,
      .f = ~ calculate_var_explained(
        data_frame = .x,
        principal_component = "value",
        attributes = c("log_age", "species_num", "sex_num")))) |>
  left_join(pca_var_df, 
            by = "PC") |>
  unnest(cols = c(var_explained)) |>
  dplyr::select(-data)
```

Since sex is mainly represented in the principal components with the least variance explained, and we don't want these to skew our further analysis, we filter our data for principal components with a variance explained over 1%.

```{r}
pca_thres_df <- pca_results_df |>
  filter(variance_explained >= 1)
```

Repeat the above analysis for the filtered data.

```{r}
var_explained_pc <- pca_thres_df |>
  group_by(PC) |>
  nest() |>
  mutate(
    var_explained = map(
      .x = data,
      .f = ~ calculate_var_explained(
        data_frame = .x,
        principal_component = "value",
        attributes = c("log_age",
                       "species_num",
                       "sex_num")))) |>
  unnest(cols = c(var_explained)) |>
  dplyr::select(-data)
```

Apply analysis to each species.

```{r}
var_explained_human <- pca_thres_df |>
  apply_variance_explained(spec = "Human",
                           attr = c("log_age",
                                    "sex_num"))

var_explained_chimp <- pca_thres_df |>
  apply_variance_explained(spec = "Chimpanzee",
                           attr = c("log_age",
                                    "sex_num"))

var_explained_macaque <- pca_thres_df |>
  apply_variance_explained(spec = "Macaque",
                           attr = c("log_age", "sex_num"))
```

Calculate PVE of sex and age for each species.

```{r}
pve_human <- var_explained_human |>
  calculate_pve() |>
  mutate(Species = "Human")

pve_chimp <- var_explained_chimp |>
  calculate_pve() |>
  mutate(Species = "Chimpanzee")

pve_macaque <- var_explained_macaque |>
  calculate_pve() |>
  mutate(Species = "Macaque")

pve_summary <- bind_rows(pve_human,
                         pve_chimp,
                         pve_macaque)
```

#### Save

Write results to .*tsv* files.

```{r}
pca_proj_df |>
  write_tsv(file = file.path(processed_data_dir,
                             "05.1_dat_pca.tsv"))
var_explained_pc_all |>
  write_tsv(file = file.path(processed_data_dir,
                             "05.2_dat_pve_pc.tsv"))
pve_summary |>
  write_tsv(file = file.path(processed_data_dir,
                             "05.3_dat_pve.tsv"))
```
