==> R/09_plot_specs.qmd <==
---
title: "Load Data"
output: html
execute:
  echo: true
  output: false
---

### Load

This document contains the specifications used for the plots of this project. We use the `RColorBrewer` package to define a set of colors.

We begin with loading the necessary libraries.

```{r}
library("tidyverse")
library("RColorBrewer")
```

#### Base colors

Our data contains two main groups of variables, 'species' and 'attributes', and one alternating variable that requires distinction by color. We use a tricolor palette consisting of the secondary colors purple, green, and orange.

Within each secondary color, a spectrum is defined to vary colors within groups.

```{r}
species_base_colors <- colorRampPalette(brewer.pal(9,
                                                   "Greens"))(3)

species_colors <- rgb(colorRamp(species_base_colors)
                      (seq(0.4, 1,
                           length.out = length(species_base_colors))),
                      maxColorValue = 255)
  
attribute_base_colors <- colorRampPalette(brewer.pal(9,
                                                     "Purples"))(3)

attribute_colors <- rgb(colorRamp(attribute_base_colors)
                      (seq(0.4, 1,
                           length.out = length(attribute_base_colors))),
                      maxColorValue = 255)

alternative_base_colors <- colorRampPalette(brewer.pal(9,
                                                       "Oranges"))(4)

alternative_colors <- rgb(colorRamp(alternative_base_colors)
                      (seq(0.3, 0.9,
                           length.out = length(alternative_base_colors))),
                      maxColorValue = 255)
```

Create data frame with colors according to group labels.

```{r}
combined_colors <- c(species_colors, attribute_colors, alternative_colors)
group_labels <- c(
  rep("Species", length(species_colors)),
  rep("Attributes", length(attribute_colors)),
  rep("Alternative", length(alternative_colors)))

color_palette <- data.frame(group = group_labels,
                            color = combined_colors)
```

#### Example

Example plot using the color palette to assign a category to `Subgroup`.

```{r}
example <- data.frame(
  group = rep(c("Species", "Attribute", "Alternative"), each = 3),
  subgroup = c("A", "B", "C", "D", "E", "F", "G", "H", "I"),
  value = c(10, 15, 20, 12, 18, 25, 8, 14, 22))

example |>
  ggplot(aes(x = group,
             y = value,
             fill = subgroup)) +
  geom_bar(stat = "identity", position = "dodge") +
  labs(
    x = "Group",
    y = "Value",
    fill = "Group",
    title = "Grouped Bar Plot with Custom Colors"
  ) +
  theme_minimal() +
  theme(
    panel.grid.major.x = element_blank(),
    panel.grid.minor.x = element_blank(),
    axis.text.x = element_text(angle = 45, hjust = 1)) +
    scale_fill_manual(values = color_palette |>
                        pull(color))
```

#### Save

Write data to *.tsv* file.

```{r}
# Set path for processed data directory
processed_data_dir <- file.path("..","data")

color_palette |>
  write_tsv(file = file.path(processed_data_dir,
                             "09_col_pal.tsv"))

```
