==> R/10_plot_PCA.qmd <==
---
title: "Principle Component Analysis, plots"
output: html
execute:
  echo: true
  output: false
---

#### Load

Load necessary libraries for plot formation including function for plotting projection data.

```{r}
library("tidyverse")
library("broom")
library("patchwork")
library("cowplot")
library("ggtext")

source("99_proj_func.R")
```

Load the data sets.

```{r}
# Set path for processed data directory
processed_data_dir <- file.path("..","data")

# Load the data sets
proj_data <- read_tsv(file = file.path(processed_data_dir,
                                         "05.1_dat_pca.tsv"),
                      show_col_types = FALSE)
pve_data <- read_tsv(file = file.path(processed_data_dir,
                                         "05.2_dat_pve_pc.tsv"),
                     show_col_types = FALSE)
color_palette <- read_tsv(file = file.path(processed_data_dir,
                                         "09_col_pal.tsv"),
                     show_col_types = FALSE)
```

#### Prepare

Define the colors used for the plots.

```{r}
age_color <- color_palette |>
  filter(group == "Attributes") |>
  slice(1) |>
  pull(color)
sex_color <- color_palette |>
  filter(group == "Attributes") |>
  slice(2) |>
  pull(color)
species_color <- color_palette |>
  filter(group == "Attributes") |>
  slice(3) |>
  pull(color)
```

The PVE data is formatted for easier plotting,

```{r}
pve_data_prep <- pve_data |>
  group_by(PC) |>
  nest() |>
  mutate(sum_var_expl_pc = map_dbl(.x = data,
                                   .f = ~ .x |>
                                     pull(var_expl_pc)|>
                                     sum())) |>
  unnest(cols = c(data)) |>
  ungroup() |>
  mutate(plt_var_expl = (var_expl_pc / sum_var_expl_pc) * variance_explained) |>
  mutate(attribute = case_when(attribute == "log_age" ~ "Age",
                            attribute == "species_num" ~ "Species",
                            attribute == "sex_num" ~ "Sex")) |>
  mutate(PC = factor(PC,
                     levels = unique(PC)))

```

### Plots

Plot the VE for each PC, colored by the PVE by attributes.

```{r}
plot_pc_pve <- pve_data_prep |>
  ggplot(aes(x = PC,
             y = plt_var_expl,
             fill = attribute)) + 
  geom_bar(stat = "identity",
           position = "stack") +
  geom_hline(yintercept = 1,
             linetype = "dashed") +
  labs(
    x = "Principal Component",
    y = "Variance Explained (%)") +
  theme_minimal() +
  theme(
    legend.position = "none",
    panel.grid.major.x = element_blank(),
    panel.grid.minor.x = element_blank(),
    panel.grid.major.y = element_blank(),
    panel.grid.minor.y = element_blank(),
    axis.text.x = element_blank(),
    axis.title = element_text(size = 12,
                              face = "italic")) +
  scale_fill_manual(values = color_palette |>
                      filter(group == "Attributes") |>
                      pull(color))

plot_pc_pve
```

Plot the projection data onto two principal components, along with details about age, species, and sex.

```{r}
# Create projection plot of PCs especially influenced by a certain attribute
plot_1_3_proj <- proj_data |>
  pca_proj_plot(principal_component_1 = "PC1",
                principal_component_2 = "PC3")

plot_2_3_proj <- proj_data |>
  pca_proj_plot(principal_component_1 = "PC2",
               principal_component_2 = "PC3")

plot_6_9_proj <- proj_data |>
  pca_proj_plot(principal_component_1 = "PC6",
                principal_component_2 = "PC9")

# Combine projection plots
proj_plots <- plot_grid(
  plot_1_3_proj,
  plot_2_3_proj,
  plot_6_9_proj,
  nrow = 1,
  rel_widths = c(1, 1, 1),
  align = "hv") |>
  ggdraw() +
  theme(plot.background = element_rect(fill = "#F5F5F5",
                                       color = NA),
        plot.margin = margin(0,0,0,0))
```

Combine plots to reveal trend of attribute influence on PC VE.

```{r}
# Add projection plots as inset element
combined_plot <- plot_pc_pve +
  inset_element(
    proj_plots,
    left = 0.1,
    right = 1,
    top = 0.9,
    bottom = 0.4) +
  
  plot_annotation(
    title = "PCA reveals trends of expression's dependencies on attributes",
    theme = theme(
      plot.title = element_text(size = 20,
                                hjust = 0,
                                face = "bold"),
      plot.margin = margin(10, 10, 10, 10))) +
  
  geom_richtext(
    aes(x = -0.01,
        y = 1.1),
    label = paste(
      "The main PVE is by the attributes <span style='color:",
      age_color,
      ";'>age</span> and ",
      "<span style='color:",
      species_color,
      ";'>species</span>..."),
    hjust = 0,
    size = 5,
    label.color = NA,
    fill = NA) +
  
  geom_richtext(
    aes(x = 1.01,
        y = -0.1),
    label = paste(
      "...whereas <span style='color:",
      sex_color,
      ";'>sex</span> is mainly represented in the PCs with lower influence."),
    hjust = 1,
    size = 5,
    label.color = NA,
    fill = NA)



combined_plot
```

#### Save

Write final plot to *.png* file.

```{r}
ggsave(file = file.path(processed_data_dir,
                        "10_PCA.png"),
       plot = combined_plot,
       width = 12,
       height = 8)
```
