==> R/11_plot_global_expr.qmd <==
---
title: "Global expression, plots"
output: html
execute:
  echo: true
  output: false
---

#### Load

Load necessary libraries for plot formation.

```{r}
library("tidyverse")
library("broom")
library("patchwork")
library("gridtext")
library("ggtext")
```

Load the data sets.

```{r}
# Set path for processed data directory
processed_data_dir <- file.path("..","data")

# Load the data set and color palette
expr_data <- read_tsv(file = file.path(processed_data_dir,
                                         "07.1_dat_global_expr.tsv"),
                      show_col_types = FALSE)
color_palette <- read_tsv(file = file.path(processed_data_dir,
                                         "09_col_pal.tsv"),
                     show_col_types = FALSE)
```

#### Prepare

Define the colors used for the plots.

```{r}
chimp_color <- color_palette |>
  filter(group == "Species") |>
  slice(1) |>
  pull(color)
hum_color <- color_palette |>
  filter(group == "Species") |>
  slice(2) |>
  pull(color)
mac_color <- color_palette |>
  filter(group == "Species") |>
  slice(3) |>
  pull(color)
```

#### Plots

The global gene expression for each sample is plotted, color-coded by species.

A sexual maturity line is added for each of the given ages:

-   Human: 16.2 years from [Human Development: A Life-Span View](https://books.google.dk/books?id=E-n5E7oyCgoC&pg=PA296&redir_esc=y)

-   Chimpanzee: 7 years days from [WWF](https://wwf.panda.org/discover/knowledge_hub/endangered_species/great_apes/chimpanzees/#:~:text=Although%20chimpanzees%20reach%20sexual%20maturity,until%20they%20are%20over%2060.)

-   Macaque: 4.9 years days from [NIH](https://pmc.ncbi.nlm.nih.gov/articles/PMC10092073/#:~:text=Cynomolgus%20macaque%20sexual%20maturation%20is,for%20CMs%20based%20on%20different)

```{r}
plot_expr_all <- expr_data |>
  ggplot(map = aes(x = age,
                   y = percent_global,
                   color = species)) +
  geom_point(size = 2) +
  geom_smooth(
    data = expr_data |>
      filter(species %in% c("Chimpanzee", "Human")),
    method = "lm",
    formula = y ~ poly(log2(x + 1), 3),
    se = FALSE) +
  # Chimpanzee
  geom_vline(xintercept = 7,
             linewidth = 0.75,
             linetype = "dashed",
             color = chimp_color) +
  # Human
  geom_vline(xintercept = 16.2,
             linewidth = 0.75,
             linetype = "dashed",
             color = hum_color) +
  # Macaque
  geom_vline(xintercept = 4.9,
             linewidth = 0.75,
             linetype = "dashed",
             color = mac_color) +
  geom_hline(yintercept = 60,
             linewidth = 0.5,
             linetype = "dashed",
             color = "gray") +
  labs(
    title = NULL,
    x = "Age (years)",
    y = "Global Expression (%)") +
  scale_x_continuous(limits = c(0, 50)) +
  scale_y_continuous(limits = c(0, 100)) +
  theme_minimal() +
  theme(legend.position = "none",
      panel.grid.major = element_blank(),
      panel.grid.minor = element_blank(),
      axis.title = element_text(size = 12,
                                face = "italic"),
      axis.line = element_line(linewidth = 0.5,
                               color = "gray"),
      axis.ticks = element_line(color = "gray")) +
  scale_color_manual(values = color_palette |>
                      filter(group == "Species") |>
                      pull(color))


plot_expr_all
```

```{r}
plot_expr_mac <- expr_data |>
  ggplot(map = aes(x = age,
                   y = percent_global,
                   color = species)) +
  geom_point(size = 2) +
  geom_smooth(
    data = expr_data |>
      filter(species %in% c("Macaque")),
    method = "lm",
    formula = y ~ log2(x + 1),
    se = FALSE,
    fullrange = TRUE) +
  # Chimpanzee
  geom_vline(xintercept = 7,
             linewidth = 0.75,
             linetype = "dashed",
             color = chimp_color) +
  # Human
  geom_vline(xintercept = 16.2,
             linewidth = 0.75,
             linetype = "dashed",
             color = hum_color) +
  # Macaque
  geom_vline(xintercept = 4.9,
             linewidth = 0.75,
             linetype = "dashed",
             color = mac_color) +
  geom_hline(yintercept = 60,
             linewidth = 0.5,
             linetype = "dashed",
             color = "gray") +
  labs(
    title = NULL,
    x = "Age (years)",
    y = NULL) +
  scale_x_continuous(limits = c(0, 20)) +
  scale_y_continuous(limits = c(0, 100)) +
  theme_minimal() +
  theme(legend.position = "none",
      panel.grid.major = element_blank(),
      panel.grid.minor = element_blank(),
      axis.title.x = element_text(size = 12,
                                 face = "italic"),
      axis.text.y = element_blank(),
      axis.line.y = element_blank(),
      axis.line.x = element_line(linewidth = 0.5,
                               color = "gray"),
      axis.ticks.x = element_line(color = "gray")) +
  scale_color_manual(values = color_palette |>
                      filter(group == "Species") |>
                      pull(color))

plot_expr_mac
```

Combine plots.

```{r}
combined_plot <- wrap_elements(
  full =
  gridtext::richtext_grob(
    text = paste0(
      "Expression reaches 60% at sexual maturity for <span style='color:",
      chimp_color,
      ";'>chimpanzee</span> and <span style='color:",
      hum_color,
      ";'>human</span>."),
    x = 0,
    hjust = 0,
    gp = grid::gpar(fontsize = 16))) /
  
  gridtext::richtext_grob(
    text = paste0(
      "Applying a log2 fit to <span style='color:",
      mac_color,
      ";'>macaque</span>, ",
      "due to reduced data points, reveals a similar trend."),
    x = 1,
    hjust = 1.05,
    gp = grid::gpar(fontsize = 16)) /
  
  (plot_expr_all + plot_expr_mac +  plot_layout(axis_titles = "collect",
                                                widths = c(4, 3))) +
  
  plot_layout(heights = c(1, 1, 8)) +
  
  plot_annotation(
     title = "Global expression over age follows the same pattern across species",
     theme = theme(
       plot.title = element_text(size = 20,
                                 hjust = 0,
                                 face = "bold")))
  


combined_plot
```

#### Save

Write final plot to *.png* file.

```{r}
ggsave(file = file.path(processed_data_dir,
                        "11_glob_expr.png"),
       plot = combined_plot,
       width = 12,
       height = 8)
```
