==> R/11_plot_shift_model.qmd <==
---
title: "Shift model, plots"
output: html
output-dir: results
execute:
  echo: true
  output: false
---

For the...

```{r}
library("tidyverse")
library("broom")
library("patchwork")
```

We load the data sets.

```{r}
# Set path for processed data directory
processed_data_dir <- file.path("..","data")

# Load the data sets
shift_data <- read_tsv(file = file.path(processed_data_dir,
                                         "07.3_hum_chi_shift.tsv"),
                     show_col_types = FALSE)
color_palette <- read_tsv(file = file.path(processed_data_dir,
                                         "04_col_pal.tsv"),
                     show_col_types = FALSE)
```

```{r}
modeled_expr = intercept +
  poly1 * log_age +
  poly2 * log_age^2 +
  poly3 * log_age^3
```

Write final plot to *.png* file.

```{r}
# ggsave(file = file.path(processed_data_dir,
#                         "10_test.png"),
#        plot = combined,
#        width = 6,
#        height = 4,
#        dpi = 300)
```
