==> R/12_plot_pve.qmd <==
---
title: "Proportion of variance explained, plots"
output: html
execute:
  echo: true
  output: false
---

### Load

Load necessary libraries for plot formation.

```{r}
library("tidyverse")
library("broom")
library("patchwork")
library("gridtext")
library("ggtext")
```

Load the data sets.

```{r}
# Set path for processed data directory
processed_data_dir <- file.path("..","data")

# Load the data sets

pve_data <- read_tsv(file = file.path(processed_data_dir,
                                      "05.3_dat_pve.tsv"),
                     show_col_types = FALSE)
mlc_proportions <- read_tsv(file = file.path(processed_data_dir,
                                           "06.2_dat_mlc_proportions.tsv"),
                          show_col_types = FALSE)
color_palette <- read_tsv(file = file.path(processed_data_dir,
                                         "09_col_pal.tsv"),
                     show_col_types = FALSE)

```

#### Prepare

Define the colors used for the plots.

```{r}
attr_color <- color_palette |>
  filter(group == "Attributes") |>
  pull(color)
age_color <- color_palette |>
  filter(group == "Attributes") |>
  slice(1) |>
  pull(color)
sex_color <- color_palette |>
  filter(group == "Attributes") |>
  slice(2) |>
  pull(color)
altern_color <- color_palette |>
  filter(group == "Alternative") |>
  pull(color)
null_color <- color_palette |>
  filter(group == "Alternative") |>
  slice(1) |>
  pull(color) 
s_color<- color_palette |>
  filter(group == "Alternative") |>
  slice(2) |>
  pull(color)
a_color<- color_palette |>
  filter(group == "Alternative") |>
  slice(3) |>
  pull(color)
as_color <- color_palette |>
  filter(group == "Alternative") |>
  slice(4) |>
  pull(color)
```

#### Plots

Donut plots for significant genes within categories.

```{r}

# Create a donut chart for human vs chimp comparison
hum_chimp_donut <- mlc_proportions |>
  ggplot(aes(x = 2, 
             y = human_chimp, 
             fill = category)) +
  geom_bar(stat = "identity", 
           width = 1, 
           color = "black") +
  coord_polar(theta = "y") +
  theme_void() +
  xlim(0.5, 2.5) +
  geom_text(aes(label = paste0(round(human_chimp),
                               "%")), 
            position = position_stack(vjust = 0.5),
            size = 4,
            color = "black",
            face = "italic") +
  scale_fill_manual(values = altern_color,
                    name = NULL) +
  theme(legend.position = "none")

# Create a donut chart for human vs macaque comparison
hum_mac_donut <- mlc_proportions |>
  ggplot(aes(x = 2, 
             y = human_macaque, 
             fill = category)) +
  geom_bar(stat = "identity", 
           width = 1, 
           color = "black") +
  coord_polar(theta = "y") +
  theme_void() +
  xlim(0.5, 2.5) +
  geom_text(aes(label = paste0(round(human_macaque),
                               "%")), 
            position = position_stack(vjust = 0.5),
            size = 4,
            color = "black",
            face = "italic") +
  scale_fill_manual(values = altern_color,
                    name = NULL) +
  theme(legend.position = "none")


# Combine donuts
donut_combined <- hum_chimp_donut + hum_mac_donut +
  plot_layout(ncol = 2) +
  plot_annotation()

hum_chimp_donut
hum_mac_donut
donut_combined

```

Bar plot for pve.

```{r}
pve_plot <- pve_data |>
  ggplot(aes(x = Species,
             y = mean_pve * 100,
             fill = attribute)) +
  geom_bar(stat = "identity",
           position = "dodge",
           show.legend = FALSE,
           color = "black") +
  labs(title = NULL,
       x = NULL,
       y = NULL) +
  theme_minimal() +
  theme(
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank()) +
  ylim(0, 90) +
  scale_y_continuous(breaks = 50,
                     limits = c(0, 90),
                     labels = function(x) paste0(x, "%")) +
  scale_fill_manual(values = attr_color) +
  geom_hline(yintercept = 50,
             linetype = "dashed",
             color = "gray",
             linewidth = 0.5) +
  coord_flip() +
  theme(
    axis.title.y = element_blank(),
    axis.text.y = element_text(vjust = -1.5,
                               size = 12,
                               face = "italic"),
    plot.margin = margin(0, 0, 0, 0))

pve_plot
```

Combine plots.

```{r}
combined_plot <- wrap_elements(
  full = 
    gridtext::richtext_grob(
      text = paste0(
        "In human <span style='color:",
        age_color,
        ";'>age</span> constiutes a bigger proportion of explained variance
        <br>
        than <span style='color:",
        sex_color,
        ";'>sex</span> compared to the other chimpanzee and macaque, "),
      x = 0,
      hjust = -0.05,
      gp = grid::gpar(fontsize = 16))) /
  
  (pve_plot + 
     
     (wrap_elements(
       full = gridtext::richtext_grob(
         text = paste0(
           "which appear to share similar patterns in the
           <br>
           percent of genes significantly dependent on
           <br>
           <span style='color:",
           a_color,
           ";'>age</span>, <span style='color:",
           s_color,
           ";'>species</span>, <span style='color:",
           as_color,
           ";'>both</span>, or <span style='color:",
           null_color,
           ";'>neither</span>."),
         x = 0.5,
         y = 0.4,
         hjust = 0.5,
         gp = grid::gpar(fontsize = 16))) /
  
      donut_combined + plot_layout(heights = c(2, 8))) +
     
     plot_layout(widths = c(2, 3))) +
  
  plot_layout(heights = c(1, 9)) +
  plot_annotation(
     title =
       "PVE by attributes appear similar in Chimpanzee and Macaque",
     theme = theme(
       plot.title = element_text(size = 20,
                                 hjust = 0,
                                 face = "bold")))
   

combined_plot
```

#### Save

Write final plot to *.png* file.

```{r}
ggsave(file = file.path(processed_data_dir,
                        "12_expr_attr.png"),
       plot = combined_plot,
       width = 12,
       height = 8)
```
